requirements.txt
uvicorn==0.18.2
fastapi==0.75.0
pystray==0.17.3
pillow==9.2.0
requests==2.26.0
pytz==2021.3
pydantic==1.9.1
PyYAML==6.0.1
